+++ R/explore.R
@@ -149,14 +149,20 @@ explore <- function(
   plot.detections.by = c("stations", "arrays")) 
 {
 
+# clean up any lost helpers
   deleteHelpers()
+  if (file.exists(paste0(tempdir(), "/actel_debug_file.txt")))
+    file.remove(paste0(tempdir(), "/actel_debug_file.txt"))
+# ------------------------
 
+# debug lines
   if (!is.null(options("actel.debug")[[1]]) && options("actel.debug")[[1]]) { # nocov start
-    on.exit(message("Debug: Progress log available at ", paste0(tempdir(), "/actel_debug_file.txt")))
-    on.exit(message("Debug: Saving carbon copy to ", paste0(tempdir(), "/actel.debug.RData")))
+    on.exit(message("Debug: Progress log available at ", gsub("\\\\", "/", paste0(tempdir(), "/actel_debug_file.txt"))))
+    on.exit(message("Debug: Saving carbon copy to ", gsub("\\\\", "/", paste0(tempdir(), "/actel.debug.RData"))), add = TRUE)
     on.exit(save(list = ls(), file = paste0(tempdir(), "/actel.debug.RData")), add = TRUE)
     message("!!!--- Debug mode has been activated ---!!!")
   } # nocov end
+# ------------------------
 
 # check arguments quality
   if (!is.null(datapack))
@@ -195,13 +201,6 @@ explore <- function(
   GUI <- checkGUI(GUI)
 # ------------------------
 
-# Prepare clean-up before function ends
-  if (file.exists(paste0(tempdir(), "/actel_debug_file.txt")))
-    file.remove(paste0(tempdir(), "/actel_debug_file.txt"))
-  on.exit(deleteHelpers(), add = TRUE)
-  on.exit(tryCatch(sink(), warning = function(w) {hide <- NA}), add = TRUE)
-# --------------------------------------
-
 # Store function call
   the.function.call <- paste0("explore(tz = ", ifelse(is.null(tz), "NULL", paste0("'", tz, "'")),
     ", datapack = ", ifelse(is.null(datapack), "NULL", deparse(substitute(datapack))),
@@ -226,7 +225,17 @@ explore <- function(
     ", print.releases = ", ifelse(print.releases, "TRUE", "FALSE"),
     ", plot.detections.by = ", plot.detections.by,
     ")")
+
+  appendTo("debug", the.function.call)
 # --------------------
+
+# Prepare clean-up before function ends
+  finished.unexpectedly <- TRUE
+  on.exit({if (finished.unexpectedly) emergencyBreak(the.function.call)}, add = TRUE)
+
+  on.exit(deleteHelpers(), add = TRUE)
+  on.exit(tryCatch(sink(), warning = function(w) {hide <- NA}), add = TRUE)
+# --------------------------------------
 
 # Final arrangements before beginning
   appendTo("Report", paste0("Actel R package report.\nVersion: ", utils::packageVersion("actel"), "\n"))